ISSUE #4906 - undo changes affected by merge
